Cleaning up and moving types into correct projects.
