+++ RapidBus.Abstractions/IIntegrationEvent.cs
@@ -3,5 +3,3 @@
 public interface IIntegrationEvent
 {
 }
-
-public delegate Task RequestDelegate(EventContext context);
