+++ RapidBus/Abstractions/Exceptions/ConfigurationException.cs
@@ -1,0 +1,7 @@
+﻿namespace RapidBus.Abstractions.Exceptions;
+
+[Serializable]
+public class ConfigurationException : RapidBusException
+{
+    internal ConfigurationException(string? message) : base(message) { }
+}
