+++ RapidBus/IRapidBus.cs
@@ -1,0 +1,9 @@
+﻿using RabidBus.Abstractions;
+
+namespace RapidBus;
+
+public interface IRapidBus
+
+{
+    void Publish<TEvent>(TEvent @event) where TEvent : IIntegrationEvent;
+}
