+++ RapidBus/Startup/StartupExtensions.cs
@@ -1,16 +1,26 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using RapidBus.Abstractions.Exceptions;
 
 namespace RapidBus;
 
 public static class StartupExtensions
 {
-    public static IServiceCollection AddRapidBus(this IServiceCollection services, Action<RapidBusOptionsBuilder> builder)
+    public static IServiceCollection AddRapidBus(this IServiceCollection services, Action<RapidBusBuilder> builder)
     {
-        var rapidBusBuilder = new RapidBusOptionsBuilder(services);
+        if (services.Any(d => d.ServiceType == typeof(IRapidBus)))
+        {
+            throw new ConfigurationException(
+                "AddRapidBus() was already called and may only be called once per container.");
+        }
+
+        //AddHostedService(services);
+        //AddInstrumentation(services);
+
+        var rapidBusBuilder = new RapidBusBuilder(services);
         builder(rapidBusBuilder);
         //services.Configure(configure);
         //services.AddSingleton<IRapidBus, RapidBus>();
-        services.AddSingleton<IEventBus>((sp) => rapidBusBuilder.Build(sp));
+        services.AddSingleton<IRapidBus>((sp) => rapidBusBuilder.Build(sp));
 
         return services;
     }
