+++ samples/RapidBus.Sample/Program.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.Hosting;
 
 using RapidBus;
+using RapidBus.Middleware;
 using RapidBus.RabbitMQ;
 using RapidBus.Sample.Event;
 using RapidBus.Sample.Middleware;
@@ -9,19 +10,26 @@ using RapidBus.Sample.Middleware;
 var builder = Host.CreateApplicationBuilder();
 
 builder.Services.AddRapidBus(builder => {
-
-    // use middleware
-    builder.UseMiddleware<SampleMiddleware>();
-
-    // use rabbitmq
-    builder.UseRabbitMQ("amqp://guest:guest@localhost:5672");
-
-    // auto register
-    builder.RegisterEventHandlers(typeof(Program).Assembly);
+    builder
+        .UseRabbitMQ("amqp://guest:guest@localhost:5672", "exchange", "queue") // use rabbitmq
+        .RegisterEventHandlers(typeof(Program).Assembly); // auto register (event processing)
 });
 
 var host = builder.Build();
-await host.RunAsync();
 
-var bus = host.Services.GetRequiredService<IEventBus>();
-bus.Publish(new SampleEvent());
+// use middleware for event processing
+host.UseEventMiddleware<SampleMiddleware>();
+
+host.Start();
+//await host.RunAsync();
+
+// sample
+
+var bus = host.Services.GetRequiredService<IRapidBus>();
+
+for (int i = 0; i < 10; i++)
+{
+    bus.Publish(new SampleEvent());
+}
+
+await Task.Delay(1999);
